requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
seaborn
scipy
Pillow
sdv
openpyxl
xlrd
torch
faker
rdt
imbalanced-learn
tabulate
